This is a fix for colons in the password of a HTTP Basic Authentication.
The existing code was taking the first component of the colon-separated array as the user name,
and the second as the password. In fact, all components after the first must be treated as belonging
to the password, and joined together with a colon separator.

Fix for https://github.com/IBM-Swift/Kitura-CredentialsHTTP/issues/54